requirements.txt
# FPS Training Simulation - Dependencies
# Phase 1: 3D Camera and Environment Setup

pygame>=2.5.0
PyOpenGL>=3.1.7
numpy>=1.24.0
